refactor : hook 파일 확장자 수정
