+++ src/refactoring/hooks/index.ts
@@ -1,3 +1,6 @@
-export * from "./useCart.ts";
-export * from "./useCoupon.ts";
-export * from "./useProduct.ts";
+export * from './useCart.ts';
+export * from './useCoupon.ts';
+export * from './useProduct.ts';
+export * from './useCouponForm.ts';
+export * from './useProductForm.ts';
+export * from './useProductEditor.ts';
